+++ packages/lab/src/index.ts
@@ -43,6 +43,11 @@ export {
   FormFieldControlWrapper,
 } from "./form-field-next";
 export * from "./form-field-context";
+export {
+  type A11yValueProps as A11yValuePropsNext,
+  type FormFieldContextNextValue,
+  FormFieldContextNext,
+} from "./form-field-context-next";
 export * from "./form-group";
 export * from "./formatted-input";
 export * from "./input";