+++ packages/lab/src/tabs/Tabstrip.tsx
@@ -107,7 +107,7 @@ export const Tabstrip = forwardRef(function Tabstrip(
   const root = useRef<HTMLDivElement>(null);
   // can't use forwardedRef here, can we ?
   // const setForkRef = useForkRef(root, forwardedRef);
-  const activeRef = useRef<number>(
+  const activeRef = useRef<number | null>(
     activeTabIndexProp || defaultActiveTabIndex || 0
   );
 
@@ -308,7 +308,11 @@ export const Tabstrip = forwardRef(function Tabstrip(
   }, [children]);
 
   useIsomorphicLayoutEffect(() => {
-    if (focusedTabIndex !== activeTabIndex && focusedTabIndex !== -1) {
+    if (
+      activeTabIndex !== null &&
+      focusedTabIndex !== activeTabIndex &&
+      focusedTabIndex !== -1
+    ) {
       tabstripHook.focusTab(activeTabIndex);
     }
 
@@ -478,7 +482,8 @@ export const Tabstrip = forwardRef(function Tabstrip(
     }
   );
 
-  const { id: selectedTabId } = collectionHook.data[activeTabIndex];
+  const selectedTabId =
+    activeTabIndex !== null ? collectionHook.data[activeTabIndex].id : null;
 
   return (
     <div